reorganize types and imports for clarity and consistency
