+++ packages/plugin-docs-renderer/src/index.ts
@@ -1,5 +1,3 @@
 // Library exports - pure markdown parsing functions only
-// (CLI utilities like filesystem scanning are not exported)
 export { parseMarkdown } from './parser.js';
 export type { ParsedMarkdown } from './parser.js';
-export type { Manifest, Page, MarkdownFiles, Frontmatter, Heading } from './types.js';
